requirements.txt
streamlit
pandas
sqlalchemy
pandas
fpdf2
psycopg2-binary